candle class conversion
