--- Intrinio.Realtime/Options/IntervalType.cs
@@ -1,14 +1,0 @@
-namespace Intrinio.Realtime.Options;
-
-public enum IntervalType
-{
-    OneMinute = 60,
-    TwoMinute = 120,
-    ThreeMinute = 180,
-    FourMinute = 240,
-    FiveMinute = 300,
-    TenMinute = 600,
-    FifteenMinute = 900,
-    ThirtyMinute = 1800,
-    SixtyMinute = 3600
-}
