+++ Intrinio.Realtime/Options/TradeCandleStick.cs
@@ -1,75 +1,28 @@
+using System.Globalization;
+
 namespace Intrinio.Realtime.Options;
 
 using System;
 
-public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, IComparable<TradeCandleStick>
+public class TradeCandleStick : CandleStick, IEquatable<TradeCandleStick>, IComparable, IComparable<TradeCandleStick>
 {
-    private readonly string _symbol;
-    private readonly double _openTimestamp;
-    private readonly double _closeTimestamp;
-    private readonly IntervalType _interval;
+    private readonly string _contract;
     
-    public string Symbol
+    public string Contract
     {
-        get { return _symbol; }
-    }
-    public UInt32 Volume { get; set; }
-    public double High { get; set; }
-    public double Low { get; set; }
-    public double Close { get; set; }
-    public double Open { get; set; }
-    public double OpenTimestamp
-    {
-        get { return _openTimestamp; }
-    }
-    public double CloseTimestamp
-    {
-        get { return _closeTimestamp; }
-    }
-    public double FirstTimestamp { get; set; }
-    public double LastTimestamp { get; set; }
-    public bool Complete { get; set; }
-    public double Average { get; set; }
-    public double Change { get; set; }
-    public IntervalType Interval
-    {
-        get { return _interval; }
-    }
-
-    public TradeCandleStick(string symbol, UInt32 volume, double price, double openTimestamp, double closeTimestamp, IntervalType interval, double tradeTime)
-    {
-        _symbol = symbol;
-        Volume = volume;
-        High = price;
-        Low = price;
-        Close = price;
-        Open = price;
-        _openTimestamp = openTimestamp;
-        _closeTimestamp = closeTimestamp;
-        FirstTimestamp = tradeTime;
-        LastTimestamp = tradeTime;
-        Complete = false;
-        Average = price;
-        Change = 0.0;
-        _interval = interval;
+        get { return _contract; }
     }
     
-    public TradeCandleStick(string symbol, UInt32 volume, double high, double low, double closePrice, double openPrice, double openTimestamp, double closeTimestamp, double firstTimestamp, double lastTimestamp, bool complete, double average, double change, IntervalType interval)
+    public TradeCandleStick(string contract, UInt32 volume, double price, double openTimestamp, double closeTimestamp, IntervalType interval, double tradeTime)
+        :base (volume, price, openTimestamp, closeTimestamp, interval, tradeTime)
     {
-        _symbol = symbol;
-        Volume = volume;
-        High = high;
-        Low = low;
-        Close = closePrice;
-        Open = openPrice;
-        _openTimestamp = openTimestamp;
-        _closeTimestamp = closeTimestamp;
-        FirstTimestamp = firstTimestamp;
-        LastTimestamp = lastTimestamp;
-        Complete = complete;
-        Average = average;
-        Change = change;
-        _interval = interval;
+        _contract = contract;
+    }
+    
+    public TradeCandleStick(string contract, UInt32 volume, double high, double low, double closePrice, double openPrice, double openTimestamp, double closeTimestamp, double firstTimestamp, double lastTimestamp, bool complete, double average, double change, IntervalType interval)
+        :base(volume, high, low, closePrice, openPrice, openTimestamp, closeTimestamp, firstTimestamp, lastTimestamp, complete, average, change, interval)
+    {
+        _contract = contract;
     }
 
     public override bool Equals(object other)
@@ -79,7 +32,7 @@ public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, ICompa
                    (!(ReferenceEquals(other, null)))
                    && (!(ReferenceEquals(this, other)))
                    && (other is TradeCandleStick)
-                   && (Symbol.Equals(((TradeCandleStick)other).Symbol))
+                   && (Contract.Equals(((TradeCandleStick)other).Contract))
                    && (Interval.Equals(((TradeCandleStick)other).Interval))
                    && (OpenTimestamp.Equals(((TradeCandleStick)other).OpenTimestamp))
                );
@@ -87,7 +40,7 @@ public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, ICompa
 
     public override int GetHashCode()
     {
-        return Symbol.GetHashCode() ^ Interval.GetHashCode() ^ OpenTimestamp.GetHashCode();
+        return Contract.GetHashCode() ^ Interval.GetHashCode() ^ OpenTimestamp.GetHashCode();
     }
 
     public bool Equals(TradeCandleStick other)
@@ -96,7 +49,7 @@ public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, ICompa
                || (
                    (!(ReferenceEquals(other, null)))
                    && (!(ReferenceEquals(this, other)))
-                   && (Symbol.Equals(other.Symbol))
+                   && (Contract.Equals(other.Contract))
                    && (Interval.Equals(other.Interval))
                    && (OpenTimestamp.Equals(other.OpenTimestamp))
                );
@@ -112,7 +65,7 @@ public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, ICompa
                 true => 1,
                 false => (other is TradeCandleStick) switch
                 {
-                    true => Symbol.CompareTo(((TradeCandleStick)other).Symbol) switch
+                    true => Contract.CompareTo(((TradeCandleStick)other).Contract) switch
                             {
                                 < 0 => -1,
                                 > 0 => 1,
@@ -137,7 +90,7 @@ public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, ICompa
             false => Object.ReferenceEquals(other, null) switch
             {
                 true => 1,
-                false => this.Symbol.CompareTo(other.Symbol) switch
+                false => this.Contract.CompareTo(other.Contract) switch
                 {
                     < 0 => -1,
                     > 0 => 1,
@@ -154,42 +107,43 @@ public class TradeCandleStick :IEquatable<TradeCandleStick>, IComparable, ICompa
 
     public override string ToString()
     {
-        return $"TradeCandleStick (Symbol: {Symbol}, Volume: {Volume.ToString()}, High: {High.ToString("f3")}, Low: {Low.ToString("f3")}, Close: {Close.ToString("f3")}, Open: {Open.ToString("f3")}, OpenTimestamp: {OpenTimestamp.ToString("f6")}, CloseTimestamp: {CloseTimestamp.ToString("f6")}, AveragePrice: {Average.ToString("f3")}, Change: {Change.ToString("f6")}, Complete: {Complete.ToString()})";
+        return $"TradeCandleStick (Contract: {Contract}, Volume: {Volume.ToString()}, High: {High.ToString("f3")}, Low: {Low.ToString("f3")}, Close: {Close.ToString("f3")}, Open: {Open.ToString("f3")}, OpenTimestamp: {OpenTimestamp.ToString("f6")}, CloseTimestamp: {CloseTimestamp.ToString("f6")}, AveragePrice: {Average.ToString("f3")}, Change: {Change.ToString("f6")}, Complete: {Complete.ToString()})";
+    }
+    
+    public string GetUnderlyingSymbol()
+    {
+        return Contract.Substring(0, 6).TrimEnd('_');
     }
 
-    public void Merge(TradeCandleStick candle)
+    public DateTime GetExpirationDate()
     {
-        Average = ((Convert.ToDouble(Volume) * Average) + (Convert.ToDouble(candle.Volume) * candle.Average)) / (Convert.ToDouble(Volume + candle.Volume));
-        Volume += candle.Volume;
-        High = High > candle.High ? High : candle.High;
-        Low = Low < candle.Low ? Low : candle.Low;
-        Close = LastTimestamp > candle.LastTimestamp ? Close : candle.Close;
-        Open = FirstTimestamp < candle.FirstTimestamp ? Open : candle.Open;
-        FirstTimestamp = candle.FirstTimestamp < FirstTimestamp ? candle.FirstTimestamp : FirstTimestamp;
-        LastTimestamp = candle.LastTimestamp > LastTimestamp ? candle.LastTimestamp : LastTimestamp;
-        Change = (Close - Open) / Open;
-    }
-            
-    internal void Update(UInt32 volume, double price, double time)
-    {
-        Average = ((Convert.ToDouble(Volume) * Average) + (Convert.ToDouble(volume) * price)) / (Convert.ToDouble(Volume + volume));
-        Volume += volume;
-        High = price > High ? price : High;
-        Low = price < Low ? price : Low;
-        Close = time > LastTimestamp ? price : Close;
-        Open = time < FirstTimestamp ? price : Open;
-        FirstTimestamp = time < FirstTimestamp ? time : FirstTimestamp;
-        LastTimestamp = time > LastTimestamp ? time : LastTimestamp;
-        Change = (Close - Open) / Open;
+        return DateTime.ParseExact(Contract.Substring(6, 6), "yyMMdd", CultureInfo.InvariantCulture);
     }
 
-    internal void MarkComplete()
+    public bool IsCall()
     {
-        Complete = true;
+        return Contract[12] == 'C';
+    }
+    
+    public bool IsPut()
+    {
+        return Contract[12] == 'P';
     }
 
-    internal void MarkIncomplete()
+    public double GetStrikePrice()
     {
-        Complete = false;
+        const UInt32 zeroChar = (UInt32)'0';
+        
+        UInt32 whole =   ((UInt32)Contract[13] - zeroChar) * 10_000u
+                         + ((UInt32)Contract[14] - zeroChar) * 1_000u
+                         + ((UInt32)Contract[15] - zeroChar) * 100u
+                         + ((UInt32)Contract[16] - zeroChar) * 10u 
+                         + ((UInt32)Contract[17] - zeroChar) * 1u;
+        
+        double part =   Convert.ToDouble((UInt32) Contract[18] - zeroChar) * 0.1D 
+                        + Convert.ToDouble((UInt32) Contract[19] - zeroChar) * 0.01D 
+                        + Convert.ToDouble((UInt32) Contract[20] - zeroChar) * 0.001D;
+        
+        return Convert.ToDouble(whole) + part;
     }
 }
